+++ c64o/print.h
@@ -13,10 +13,10 @@
 
 void print_label(uint16_t pos, const char *label);
 void print_bcd(uint16_t pos, uint32_t value, uint8_t num_digits = 5);
-void print_signed_bcd(uint16_t pos, int32_t value, uint8_t num_digits = 5);
+void print_signed_bcd(uint16_t pos, int16_t value, uint8_t num_digits = 5);
 void print_labeled_bcd(uint16_t pos, const char *label, uint32_t value,
                        uint8_t num_digits = 5);
-void print_labeled_signed_bcd(uint16_t pos, const char *label, int32_t value,
+void print_labeled_signed_bcd(uint16_t pos, const char *label, int16_t value,
                               uint8_t num_digits = 5);
 void print_hex(uint16_t pos, uint32_t value, uint8_t num_digits = 4);
 void print_labeled_hex(uint16_t pos, const char *label, uint32_t value,
